- sharepage refactoring
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1476 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ AutoSearchManager.cpp
@@ -245,7 +245,7 @@ void AutoSearch::updateStatus() {
 		if (manualSearch) {
 			status = AutoSearch::STATUS_MANUAL;
 		} else {
-			status = (expireTime > 0 && static_cast<uint64_t>(expireTime) < GET_TIME()) ? AutoSearch::STATUS_EXPIRED : AutoSearch::STATUS_DISABLED;
+			status = (expireTime > 0 && expireTime < GET_TIME()) ? AutoSearch::STATUS_EXPIRED : AutoSearch::STATUS_DISABLED;
 		}
 		return;
 	}
